requirements.txt
click==8.1.7
gunicorn==23.0.0
Jinja2==3.1.4
parse==1.20.2
pytest==8.3.2
requests==2.32.3
requests-wsgi-adapter==0.4.1
WebOb==1.8.8
whitenoise==6.7.0
